v3: use narrower integer types in the AST and node-indexed tables (#28401)

+++ vlib/v3/parser/parser_parallel_notd_v3_no_parallel.v
@@ -318,7 +318,7 @@ fn remap_worker_pos(pos token.Pos, first_file_id int, next_file_id int, delta in
 	}
 	return token.Pos{
 		...pos
-		id: pos.id + delta
+		id: i32(pos.id + delta)
 	}
 }
